renamed test.js -> mealRoutes.js for clarity
